Linked notes to player health, made enemy
Made a very basic player and enemy place holder objects and scripts. BattleManager has two signals, it will send them based on score to either player or enemy. Player and enemy have a healthBar and a ref to the BattleManagerInstance. Could be cleaner with where the actual health variable is stored.