Fix Not checking misses correctly after loops
For when I go back and look at this commit:
The issue is we say songs can be longer than an exact int num of loops.
So a song can eventually have a realbeat > beatsperloop.

The current logic is a bit blind, and:
(if the realbeat is ahead of last beat) -> increment last beat to catch up.

But last beat is mod beatsperloop, so it tries to catch up while realbeat is a float beyond beatsperloop, so it keeps overrunning until we fully loop and realbeat is back to approximately 0.

This fix just continues if realbeat is beyond beatsperloop.

This does not fix the underlying issue that lastbeat can absolutely be desynced from the actual beat, because I'm already up too late.