Numbers balancing (#232)
* Enemy money changes

standardized the money from fights based on enemy tier

* nerfed blood money, other relics and notes look fine

* Adjusted enemy HP

Only strange value to me is Parasifly, since we use one template and the fight can either be a 1v1 or 1v2, but this shouldn't be too big of an issue.

* song speed changes

value of 200 for all songs, notes now take about 2 seconds to go from being visible to being hit

* standardized battle countdown

* Undo countdown standardizing

This was intentionally mis-timed to signify the BPM of a song.

* Price adjustments

* Adjust heal cost translation

---------

Co-authored-by: LifeHckr <jarodthereal@gmail.com>