Added relic/note/money to rob text (#233)
* Added relic/note/money to rob text

The text should display the name of the relic/note/money taken from the player in paraenthesis.  EventTemplate was slightly tweaked to include extra func OutcomeResolver. Adjusted EventDatabase and Event Scene accordingly.

* Drastic Simplification

One shortcoming with the and the previous version is if locale is changed while these messages are displayed they don't auto change.
Doesn't seem to be a good way to accomplish this regardless.

* Added support for the wheel results

Should work correctly

* Finalize Event 2 Text

---------

Co-authored-by: LifeHckr <jarodthereal@gmail.com>